EU-115 refactor services
