+++ apps/services/handlers.py
@@ -1,62 +1,42 @@
 from apps.common.db import database
 from apps.services.models import user_request_exist_view, user_request_list_view, STATUS_MAPPING, UserRequest, \
     user_request_booking_hostel_view, UserRequestReview, hostel_accommodation_view, user_request_details_view
-from apps.services.schemas import UserRequestExistenceOut, UserRequestsListOut, CreateUserRequestOut, \
-    CreateUserRequestIn, UserRequestBookingHostelOut, CancelRequestOut, CancelRequestIn, UserRequestReviewOut, \
-    UserRequestReviewIn, HostelAccomodationViewOut, UserRequestDetailsViewOut
+from apps.services.schemas import CreateUserRequestIn, CancelRequestIn, UserRequestReviewIn
 from apps.services.services import create_user_document
-from apps.users.handlers import get_current_user
 from apps.users.models import UserFaculty
+from apps.users.schemas import UserOut
 
 from datetime import datetime
-from typing import List
 import json
 
-from fastapi import Depends, APIRouter
 from sqlalchemy import select, insert, update
-from apps.common.schemas import JSENDOutSchema
-services_router = APIRouter()
 
 
-@services_router.get("/{university_id}/user-request-existence/{service_id}/",
-            response_model=JSENDOutSchema[UserRequestExistenceOut], tags=["Student dashboard"])
-async def check_user_request_existence(university_id: int, service_id: int, user=Depends(get_current_user)):
+async def get_user_request_existence(university_id: int, service_id: int, user: UserOut):
     query = select(user_request_exist_view).where(user_request_exist_view.c.user_id == user.user_id,
                                                   user_request_exist_view.c.university_id == university_id,
                                                   user_request_exist_view.c.service_id == service_id)
     user_request_result = await database.fetch_one(query)
     if user_request_result:
-        response = {
+        return {
             "user_request_id": user_request_result.user_request_id,
             "status": json.loads(user_request_result.status),
             "user_request_exist": True
         }
-    else:
-        response = {
-            "user_request_id": None,
-            "status": None,
-            "user_request_exist": False
-        }
     return {
-        "data": response,
-        "message": "Got user request existence"
+        "user_request_id": None,
+        "status": None,
+        "user_request_exist": False
     }
 
 
-@services_router.get("/{university_id}/user-request/", response_model=JSENDOutSchema[List[UserRequestsListOut]],
-                     tags=["Student dashboard"])
-async def read_user_request_list(university_id: int, user=Depends(get_current_user)):
+async def get_user_request_list(university_id: int, user: UserOut):
     query = select(user_request_list_view).where(user_request_list_view.c.user_id == user.user_id,
                                                  user_request_list_view.c.university_id == university_id)
-    return {
-        "data": await database.fetch_all(query),
-        "message": "Got user requests list"
-    }
+    return await database.fetch_all(query)
 
 
-@services_router.post("/{university_id}/user-request/", response_model=JSENDOutSchema[CreateUserRequestOut],
-                      tags=["Student dashboard"])
-async def create_user_request(university_id: int, user_request: CreateUserRequestIn, user=Depends(get_current_user)):
+async def post_user_request(university_id: int, user_request: CreateUserRequestIn, user: UserOut):
     query = select(UserFaculty).where(UserFaculty.user_id == user.user_id)
     user_faculty_result = await database.fetch_one(query)
     query = insert(UserRequest).values(date_created=datetime.now(),
@@ -69,8 +49,10 @@ async def create_user_request(university_id: int, user_request: CreateUserReques
 
     last_record_id = await database.execute(query)
 
-    query = select(user_request_booking_hostel_view).where(user_request_booking_hostel_view.c.user_id == user.user_id,
-                                                           user_request_booking_hostel_view.c.university_id == university_id)
+    query = select(user_request_booking_hostel_view).where(
+        user_request_booking_hostel_view.c.user_id == user.user_id,
+        user_request_booking_hostel_view.c.university_id == university_id
+    )
 
     result = await database.fetch_one(query)
 
@@ -83,48 +65,32 @@ async def create_user_request(university_id: int, user_request: CreateUserReques
     await create_user_document(**prepared_data)
 
     return {
-        "data": {
             "status_id": STATUS_MAPPING.get("Розглядається"),
             "user_request_id": last_record_id
-        },
-        "message": f"Created user request with id {last_record_id}"
-    }
+        }
 
 
-@services_router.get("/{university_id}/user-request-booking-hostel/", response_model=JSENDOutSchema[UserRequestBookingHostelOut],
-                     tags=["Student dashboard"])
-async def read_user_request_booking_hostel(university_id: int, user=Depends(get_current_user)):
-    query = select(user_request_booking_hostel_view).where(user_request_booking_hostel_view.c.user_id == user.user_id,
-                                                           user_request_booking_hostel_view.c.university_id == university_id)
-    return {
-        "data": await database.fetch_one(query),
-        "message": "Got user request booking hostel"
-    }
+async def get_user_request_booking_hostel(university_id: int, user: UserOut):
+    query = select(user_request_booking_hostel_view).where(
+        user_request_booking_hostel_view.c.user_id == user.user_id,
+        user_request_booking_hostel_view.c.university_id == university_id
+    )
+    return await database.fetch_one(query)
 
 
-@services_router.put("/{university_id}/user-request/{user_request_id}", response_model=JSENDOutSchema[CancelRequestOut],
-                     tags=["Student dashboard"])
-async def cancel_request(university_id: int, user_request_id: int, cancel_request: CancelRequestIn,
-                         user=Depends(get_current_user)):
+async def put_cancel_request(user_request_id: int, cancel_request: CancelRequestIn):
     CancelRequestIn(status_id=cancel_request.status_id)
     query = update(UserRequest).where(UserRequest.user_request_id == user_request_id).values(
         status_id=cancel_request.status_id)
     await database.execute(query)
-
     return {
-        "data": {
-            "user_request_id": user_request_id,
-            "status_id": cancel_request.status_id
-        },
-        "message": f"Canceled request with id {user_request_id}"
+        "user_request_id": user_request_id,
+        "status_id": cancel_request.status_id
     }
 
 
-@services_router.post("/{university_id}/user-request-review/{user_request_id}/",
-                      response_model=JSENDOutSchema[UserRequestReviewOut],
-                      tags=["Admin dashboard"])
-async def create_user_request_review(university_id: int, user_request_id: int, user_request_review: UserRequestReviewIn,
-                                     user=Depends(get_current_user)):
+async def post_user_request_review(university_id: int, user_request_id: int, user_request_review: UserRequestReviewIn,
+                                   user: UserOut):
     query = insert(UserRequestReview).values(university_id=university_id,
                                              user_request_id=user_request_id,
                                              date_created=datetime.now(),
@@ -141,21 +107,18 @@ async def create_user_request_review(university_id: int, user_request_id: int, u
 
     last_record_id = await database.execute(query)
 
-    query = update(UserRequest).values(status_id=user_request_review.status_id).where(UserRequest.user_request_id == user_request_id)
+    query = update(UserRequest).values(status_id=user_request_review.status_id).where(
+        UserRequest.user_request_id == user_request_id
+    )
     await database.execute(query)
 
     return {
-        "data": {
-            "status_id": user_request_review.status_id,
-            "user_request_review_id": last_record_id
-        },
-        "message": "Created user request review"
+        "status_id": user_request_review.status_id,
+        "user_request_review_id": last_record_id
     }
 
 
-@services_router.get("/{university_id}/hostel-accommodation/{user_request_id}",
-                     response_model=JSENDOutSchema[HostelAccomodationViewOut], tags=["Student dashboard"])
-async def read_hostel_accommodation(university_id: int, user_request_id: int, user=Depends(get_current_user)):
+async def get_hostel_accommodation(university_id: int, user_request_id: int):
     query = select(hostel_accommodation_view).where(hostel_accommodation_view.c.university_id == university_id,
                                                     hostel_accommodation_view.c.user_request_id == user_request_id)
     response = await database.fetch_one(query)
@@ -164,15 +127,10 @@ async def read_hostel_accommodation(university_id: int, user_request_id: int, us
 
     response.hostel_name = json.loads(response.hostel_name)
     response.hostel_address = json.loads(response.hostel_address)
-    return {
-        "data": response,
-        "message": "Get hostel accommodation"
-    }
+    return response
 
 
-@services_router.get("/{university_id}/user-request/{user_request_id}", response_model=JSENDOutSchema[UserRequestDetailsViewOut],
-                     tags=["Student dashboard"])
-async def read_request_details(university_id: int, user_request_id: int, user=Depends(get_current_user)):
+async def get_request_details(university_id: int, user_request_id: int):
     query = select(user_request_details_view).where(user_request_details_view.c.university_id == university_id,
                                                     user_request_details_view.c.user_request_id == user_request_id)
 
@@ -180,7 +138,4 @@ async def read_request_details(university_id: int, user_request_id: int, user=De
 
     response.documents = json.loads(response.documents)
     response.hostel_name = json.loads(response.hostel_name)
-    return {
-        "data": response,
-        "message": "Got request details"
-    }
+    return response
